Revert "Re-organize project files into a clean, purposeful src-layout structure"
