+++ .gitignore
@@ -5,18 +5,3 @@
 **/__pycache__/
 .mypy_cache/
 .ruff_cache/
-
-# Build artifacts
-/build/
-/dist/
-*.egg-info/
-
-# IDE files
-.vscode/
-.idea/
-*.swp
-*.swo
-
-# OS files
-.DS_Store
-Thumbs.db
