+++ dynamics.py
