+++ plotter.py
@@ -1,0 +1,3 @@
+from data_manager import results
+
+results()
