--- src/online_adaptive_resnet.egg-info/SOURCES.txt
@@ -1,26 +1,0 @@
-LICENSE
-README.md
-pyproject.toml
-src/online_adaptive_resnet/__init__.py
-src/online_adaptive_resnet/__main__.py
-src/online_adaptive_resnet.egg-info/PKG-INFO
-src/online_adaptive_resnet.egg-info/SOURCES.txt
-src/online_adaptive_resnet.egg-info/dependency_links.txt
-src/online_adaptive_resnet.egg-info/entry_points.txt
-src/online_adaptive_resnet.egg-info/requires.txt
-src/online_adaptive_resnet.egg-info/top_level.txt
-src/online_adaptive_resnet/cli/__init__.py
-src/online_adaptive_resnet/cli/main.py
-src/online_adaptive_resnet/core/__init__.py
-src/online_adaptive_resnet/core/dynamics.py
-src/online_adaptive_resnet/core/entity.py
-src/online_adaptive_resnet/core/neural_network.py
-src/online_adaptive_resnet/io/__init__.py
-src/online_adaptive_resnet/io/config.py
-src/online_adaptive_resnet/io/data_manager.py
-src/online_adaptive_resnet/utils/__init__.py
-src/online_adaptive_resnet/utils/integrate.py
-src/online_adaptive_resnet/visualization/__init__.py
-src/online_adaptive_resnet/visualization/plotter.py
-tests/test_resnet_reference.py
-tests/test_resnet_reference_backup.py
