--- src/online_adaptive_resnet.egg-info/entry_points.txt
@@ -1,2 +1,0 @@
-[console_scripts]
-online-adaptive-resnet = online_adaptive_resnet.cli.main:main
