--- src/online_adaptive_resnet.egg-info/requires.txt
@@ -1,5 +1,0 @@
-numpy
-scipy
-matplotlib
-pandas
-scienceplots
