--- src/online_adaptive_resnet.egg-info/top_level.txt
@@ -1,1 +1,0 @@
-online_adaptive_resnet
