--- src/online_adaptive_resnet/__init__.py
@@ -1,9 +1,0 @@
-"""Online Adaptive Deep Residual Neural Network.
-
-A simulation framework for online adaptive deep neural networks with residual
-connections designed for modeling and controlling dynamical systems.
-"""
-
-__version__ = "1.0.0"
-__author__ = "Cristian F. Nino"
-__email__ = "cristian1928@ufl.edu"
