--- src/online_adaptive_resnet/__main__.py
@@ -1,6 +1,0 @@
-"""Entry point for running the package as a module."""
-
-from .cli.main import main
-
-if __name__ == "__main__":
-    main()
