--- src/online_adaptive_resnet/cli/__init__.py
@@ -1,1 +1,0 @@
-"""Command-line interface."""
