--- src/online_adaptive_resnet/core/__init__.py
@@ -1,1 +1,0 @@
-"""Core neural network and entity implementations."""
