--- src/online_adaptive_resnet/io/__init__.py
@@ -1,1 +1,0 @@
-"""Input/output utilities and data management."""
