--- src/online_adaptive_resnet/visualization/__init__.py
@@ -1,1 +1,0 @@
-"""Visualization and plotting utilities."""
